+++ CHANGELOG.md
@@ -30,6 +30,9 @@ Qristal is a full-stack SDK for quantum accelerators.
 - Added maximum likelihood estimation to standard quantum state tomography protocol. Can be enabled by calling `set_maximum_likelihood_estimation()`.
 - Added process matrix solver and interpolator.
 - Enabled Kraus matrix contribution probability to improve qsim Kraus matrix sampling.
+- Added new `qristal::session` member functions `set_SPAM_confusion_matrix()` and `set_SPAM_correction_matrix()` to add a suitable state preparation and measurement (SPAM) confusion or correction matrix, respectively. The latter will be used to automatically correct measured results, overwriting the `results_` variable. Native results are kept separately in`results_native` instead. 
+- Added the python binding `qristal.core.session.SPAM_confusion` to interface with `qristal::session::set_SPAM_confusion_matrix()`, and `qristal.core.session.results_native` with 
+`qristal::session::results_native()`.
 - Added parameterized CRZ, CRX and CRY gates providing controlled Z, X and Y gates respectively. CRZ is equivalent to parameterized CPhase. Also added python wrappers crz, crx and cry.
 
 ### Fixed